v0.64.58.4

+++ MTEngine/Games/c64/Emulators/atari800/AtariInterface/AtariDebugInterface.h
@@ -60,12 +60,12 @@ public:
 	virtual void HardReset();
 	
 	// this is main emulation cpu cycle counter
-	virtual unsigned int GetMainCpuCycleCounter();
-	virtual unsigned int GetPreviousCpuInstructionCycleCounter();
+	virtual u64 GetMainCpuCycleCounter();
+	virtual u64 GetPreviousCpuInstructionCycleCounter();
 	
 	// resettable counters for debug purposes
 	virtual void ResetMainCpuDebugCycleCounter();
-	virtual unsigned int GetMainCpuDebugCycleCounter();
+	virtual u64 GetMainCpuDebugCycleCounter();
 	
 	virtual void RefreshScreenNoCallback();
